chore: refactor the Add Asset modal component
