--- static/src/components/add-asset-modal.jsx
@@ -1,5 +1,0 @@
-import { AssetModal } from './asset-modal'
-
-export const AddAssetModal = (props) => {
-  return <AssetModal {...props} />
-}
