+++ static/src/components/add-asset-modal/file-upload-utils.js
