+++ static/src/components/add-asset-modal/uri-tab.jsx
